Move types to new types.ts file
