+++ src/types.ts
@@ -1,0 +1,8 @@
+export type HeatMapArray = number[][];
+
+export type PositionArray = ({ name: string; hit: boolean } | null)[][];
+
+export type ShipInfo = {
+  name: 'carrier' | 'battleship' | 'cruiser' | 'submarine' | 'destroyer';
+  size: number;
+};
